refactor: move StringExtension and Validator to login/domain/util and add login/manager
